# output perf: avoid redundant array copies (plugin-transform-spread)