--- a/benchmark/babel-plugin-transform-spread/array-spread-arguments.mjs
+++ b/benchmark/babel-plugin-transform-spread/array-spread-arguments.mjs
@@ -1,0 +1,15 @@
+import runner from "./runner.mjs";
+
+runner({
+  code: `
+    function arrayOf() {
+      return [...arguments];
+    }
+
+    function work() {
+      return arrayOf("bacon", "egg", "spam");
+    }
+  `,
+  expect: ["bacon", "egg", "spam"],
+  loopCount: 100,
+});