--- a/benchmark/babel-plugin-transform-spread/array-spread-multiple.mjs
+++ b/benchmark/babel-plugin-transform-spread/array-spread-multiple.mjs
@@ -1,0 +1,13 @@
+import runner from "./runner.mjs";
+
+runner({
+  code: `
+    var bar = ["egg", , "spam"];
+
+    function work(a = bar) {
+      return [...a, ...a, ...a];
+    }
+  `,
+  expect: ["egg", void 0, "spam", "egg", void 0, "spam", "egg", void 0, "spam"],
+  loopCount: 100,
+});