--- a/benchmark/babel-plugin-transform-spread/call-spread-only-arguments.mjs
+++ b/benchmark/babel-plugin-transform-spread/call-spread-only-arguments.mjs
@@ -1,0 +1,19 @@
+import runner from "./runner.mjs";
+
+runner({
+  code: `
+    function foo() {
+      return arguments.length;
+    }
+
+    function bar() {
+      return foo(...arguments);
+    }
+
+    function work() {
+      return bar("bacon", "egg", "spam");
+    }
+  `,
+  expect: 3,
+  loopCount: 100,
+});