--- a/packages/babel-helpers/src/helpers/iteratorClose.js
+++ b/packages/babel-helpers/src/helpers/iteratorClose.js
@@ -1,0 +1,5 @@
+/* @minVersion 7.16.5 */
+
+export default function _iteratorClose(it) {
+  if (it.return != null) it.return();
+}