--- a/packages/babel-helpers/src/helpers/iteratorCloseQuiet.js
+++ b/packages/babel-helpers/src/helpers/iteratorCloseQuiet.js
@@ -1,0 +1,9 @@
+/* @minVersion 7.16.5 */
+
+export default function _iteratorCloseQuiet(it) {
+  try {
+    if (it.return != null) it.return();
+  } catch (e) {
+    // swallow exceptions
+  }
+}