--- a/packages/babel-plugin-proposal-class-properties/test/fixtures/regression/6154/output.js
+++ b/packages/babel-plugin-proposal-class-properties/test/fixtures/regression/6154/output.js
@@ -17,7 +17,7 @@ var Test = function Test() {
         args[_key] = arguments[_key];
       }
 
-      _this = _super.call.apply(_super, [this].concat(args));
+      _this = _super.call.apply(_super, babelHelpers.appendArrayLike([this], args));
       babelHelpers.defineProperty(babelHelpers.assertThisInitialized(_this), "a", function () {
         return babelHelpers.get((_thisSuper = babelHelpers.assertThisInitialized(_this), babelHelpers.getPrototypeOf(Other.prototype)), "test", _thisSuper);
       });