--- a/packages/babel-plugin-proposal-optional-chaining/test/fixtures/general/function-call-spread/output.js
+++ b/packages/babel-plugin-proposal-optional-chaining/test/fixtures/general/function-call-spread/output.js
@@ -1,6 +1,6 @@
 var _a, _a2, _a3, _a4, _a4$b;
 
-(_a = a) === null || _a === void 0 ? void 0 : _a.apply(void 0, babelHelpers.toConsumableArray(args));
-(_a2 = a) === null || _a2 === void 0 ? void 0 : _a2.b.apply(_a2, babelHelpers.toConsumableArray(args));
-(_a3 = a) === null || _a3 === void 0 ? void 0 : _a3.b.apply(_a3, babelHelpers.toConsumableArray(args)).c;
-(_a4 = a) === null || _a4 === void 0 ? void 0 : (_a4$b = _a4.b.apply(_a4, babelHelpers.toConsumableArray(args))).c.apply(_a4$b, babelHelpers.toConsumableArray(args));
+(_a = a) === null || _a === void 0 ? void 0 : _a.apply(void 0, babelHelpers.spreadCoerceToArray(args));
+(_a2 = a) === null || _a2 === void 0 ? void 0 : _a2.b.apply(_a2, babelHelpers.spreadCoerceToArray(args));
+(_a3 = a) === null || _a3 === void 0 ? void 0 : _a3.b.apply(_a3, babelHelpers.spreadCoerceToArray(args)).c;
+(_a4 = a) === null || _a4 === void 0 ? void 0 : (_a4$b = _a4.b.apply(_a4, babelHelpers.spreadCoerceToArray(args))).c.apply(_a4$b, babelHelpers.spreadCoerceToArray(args));