--- a/packages/babel-plugin-transform-classes/test/fixtures/assumption-constantSuper/accessing-super-class/output.js
+++ b/packages/babel-plugin-transform-classes/test/fixtures/assumption-constantSuper/accessing-super-class/output.js
@@ -6,7 +6,7 @@ var Test = /*#__PURE__*/function (_Foo) {
   var _super = babelHelpers.createSuper(Test);
 
   function Test() {
-    var _Foo$prototype$test;
+    var _sprd, _sprd2, _Foo$prototype$test;
 
     var _this;
 
@@ -17,11 +17,11 @@ var Test = /*#__PURE__*/function (_Foo) {
     _Foo.prototype.test.call(babelHelpers.assertThisInitialized(_this));
 
     _this = _super.apply(this, arguments);
-    _this = _super.call.apply(_super, [this, "test"].concat(Array.prototype.slice.call(arguments)));
+    _this = _super.call.apply(_super, ((_sprd = [this, "test"]).push.apply(_sprd, arguments), _sprd));
 
     _Foo.prototype.test.apply(babelHelpers.assertThisInitialized(_this), arguments);
 
-    (_Foo$prototype$test = _Foo.prototype.test).call.apply(_Foo$prototype$test, [babelHelpers.assertThisInitialized(_this), "test"].concat(Array.prototype.slice.call(arguments)));
+    (_Foo$prototype$test = _Foo.prototype.test).call.apply(_Foo$prototype$test, ((_sprd2 = [babelHelpers.assertThisInitialized(_this), "test"]).push.apply(_sprd2, arguments), _sprd2));
 
     return _this;
   }