--- a/packages/babel-plugin-transform-classes/test/fixtures/loose/accessing-super-class/output.js
+++ b/packages/babel-plugin-transform-classes/test/fixtures/loose/accessing-super-class/output.js
@@ -4,7 +4,7 @@ var Test = /*#__PURE__*/function (_Foo) {
   babelHelpers.inheritsLoose(Test, _Foo);
 
   function Test() {
-    var _Foo$prototype$test;
+    var _sprd, _sprd2, _Foo$prototype$test;
 
     var _this;
 
@@ -14,11 +14,11 @@ var Test = /*#__PURE__*/function (_Foo) {
     _Foo.prototype.test.call(babelHelpers.assertThisInitialized(_this));
 
     _this = _Foo.apply(this, arguments) || this;
-    _this = _Foo.call.apply(_Foo, [this, "test"].concat(Array.prototype.slice.call(arguments))) || this;
+    _this = _Foo.call.apply(_Foo, ((_sprd = [this, "test"]).push.apply(_sprd, arguments), _sprd)) || this;
 
     _Foo.prototype.test.apply(babelHelpers.assertThisInitialized(_this), arguments);
 
-    (_Foo$prototype$test = _Foo.prototype.test).call.apply(_Foo$prototype$test, [babelHelpers.assertThisInitialized(_this), "test"].concat(Array.prototype.slice.call(arguments)));
+    (_Foo$prototype$test = _Foo.prototype.test).call.apply(_Foo$prototype$test, ((_sprd2 = [babelHelpers.assertThisInitialized(_this), "test"]).push.apply(_sprd2, arguments), _sprd2));
 
     return _this;
   }