--- a/packages/babel-plugin-transform-spread/test/fixtures/assumption-arrayLikeIsIterable/simple/output.js
+++ b/packages/babel-plugin-transform-spread/test/fixtures/assumption-arrayLikeIsIterable/simple/output.js
@@ -1,1 +1,1 @@
-var arr = ["a"].concat(babelHelpers.maybeArrayLike(babelHelpers.toConsumableArray, p2), ["e"]);
+var arr = babelHelpers.concatArrayLike(["a"], babelHelpers.spreadCoerceToArrayLike(p2), ["e"]);