--- a/packages/babel-plugin-transform-spread/test/fixtures/assumption-iterableIsArray/arguments/output.js
+++ b/packages/babel-plugin-transform-spread/test/fixtures/assumption-iterableIsArray/arguments/output.js
@@ -1,5 +1,7 @@
 function foo() {
+  var _sprd;
+
   // We know for sure that 'arguments' is _not_ an array, so we
   // can ignore the assumption in this case.
-  return Array.prototype.slice.call(arguments);
+  return (_sprd = []).push.apply(_sprd, arguments), _sprd;
 }