… benchmark/package.json
… packages/babel-plugin-transform-runtime/test/fixtures/regression/12863/output.mjs
… packages/babel-plugin-transform-spread/src/index.ts
… packages/babel-plugin-transform-spread/test/fixtures/assumption-iterableIsArray/array-literals/output.js
… packages/babel-plugin-transform-spread/test/fixtures/assumption-iterableIsArray/function-call/output.js
… packages/babel-plugin-transform-spread/test/fixtures/assumption-iterableIsArray/sparse-array/exec.js
… packages/babel-plugin-transform-spread/test/fixtures/call-context/flow-type-cast/output.js
… packages/babel-plugin-transform-spread/test/fixtures/call-context/parenthesized-expressions/output.js
… packages/babel-plugin-transform-spread/test/fixtures/call-context/ts-type-cast/output.js
… packages/babel-plugin-transform-spread/test/fixtures/regression/10416/output.mjs
… packages/babel-plugin-transform-spread/test/fixtures/regression/11400/output.js
… packages/babel-plugin-transform-spread/test/fixtures/regression/6647/output.js
… packages/babel-plugin-transform-spread/test/fixtures/regression/issue-8907/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/arguments-array/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/arguments-concat/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/array-literal-first/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/array-literal-middle/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/array-literal-multiple/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/array-literal-with-hole/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/array-literals/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/array-symbol-unsupported/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/contexted-computed-method-call-multiple-args/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/contexted-computed-method-call-single-arg/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/contexted-method-call-multiple-args/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/contexted-method-call-single-arg/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/contexted-method-call-super-multiple-args/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/contexted-method-call-super-single-arg/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/known-rest/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/method-call-array-literal-babel-7/output.js
… packages/babel-plugin-transform-spread/test/fixtures/spread/method-call-array-literal/output.js