<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | Fixes #13861 
| Patch: Bug Fix?          | :heavy_check_mark: 
| Major: Breaking Change?  |
| Minor: New Feature?      | some new helpers, do those count?
| Tests Added + Pass?      | Yes
| Documentation PR Link    | <!-- If only readme change, add `[skip ci]` to your commits -->
| Any Dependency Changes?  | only internal, in benchmark devdeps
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->

### 1. bug fix: spreading sparse array variable
An accidental side effect of this revamp is that it fixes #13861. (test included)

### 2. perf: spreading `arguments`
Previously `Array.p.slice` was used to convert `[...arguments]` to array, but `Array.p.push` works faster.

```
array-spread-arguments baseline: 16_773 ops/sec ±0.41% (0.06ms)
array-spread-arguments current: 33_122 ops/sec ±0.33% (0.03ms)
```
<details>
<summary>benchmark transform</summary>

```js
// array-spread-arguments baseline
function arrayOf() {
  return Array.prototype.slice.call(arguments);
}

function work() {
  return arrayOf("bacon", "egg", "spam");
}
```
```js
// array-spread-arguments current
function arrayOf() {
  var _sprd;

  return (_sprd = []).push.apply(_sprd, arguments), _sprd;
}

function work() {
  return arrayOf("bacon", "egg", "spam");
}
```
</details>


### 3. perf: spreading single array in call
A call such as `foo(...bar)` was transformed into `foo.apply(void 0, _toConsumableArray(bar))`.
`_toConsumableArray` returns a copy without holes. But if `bar` is already an array, there is no need to make a copy, even if it is sparse: `foo.apply` does not respect holes. At first I simply replaced `_toConsumableArray` with `_toArray`, but that helper was written for destructuring and gives a misleading error message when `bar` is not iterable. So I ended up adding a new helper.

```
call-spread-sparse-array baseline: 7_221 ops/sec ±0.52% (0.138ms)
call-spread-sparse-array current: 12_735 ops/sec ±0.33% (0.079ms)
```

<details>
<summary>benchmark transform</summary>

```js
// call-spread-sparse-array baseline
function _toConsumableArray(arr) { return _arrayWithoutHoles(arr) || _iterableToArray(arr) || _unsupportedIterableToArray(arr) || _nonIterableSpread(); }

function _nonIterableSpread() { throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }

function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(o); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }

function _iterableToArray(iter) { if (typeof Symbol !== "undefined" && iter[Symbol.iterator] != null || iter["@@iterator"] != null) return Array.from(iter); }

function _arrayWithoutHoles(arr) { if (Array.isArray(arr)) return _arrayLikeToArray(arr); }

function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) arr2[i] = arr[i]; return arr2; }

// defer initialization to prevent Babel guessing variable type,
// because if it can guess, the plugin emits Array-specific code
var bar;
bar = new Array(3);

function foo() {
  return arguments.length;
}

function work() {
  return foo.apply(void 0, _toConsumableArray(bar));
}
```
```js
// call-spread-sparse-array current
function _spreadCoerceToArray(src) { return Array.isArray(src) ? src : _iterableToArray(src) || _unsupportedIterableToArray(src) || _nonIterableSpread(); }

function _nonIterableSpread() { throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }

function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(o); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }

function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) arr2[i] = arr[i]; return arr2; }

function _iterableToArray(iter) { if (typeof Symbol !== "undefined" && iter[Symbol.iterator] != null || iter["@@iterator"] != null) return Array.from(iter); }

// defer initialization to prevent Babel guessing variable type,
// because if it can guess, the plugin emits Array-specific code
var bar;
bar = new Array(3);

function foo() {
  return arguments.length;
}

function work() {
  return foo.apply(void 0, _spreadCoerceToArray(bar));
}
```
</details>

### 4. perf: calls with multiple arguments and spread
Previously calls like `foo("ham", "spam", ...arguments, "ham", "spam");` were transformed such that each spread element was copied to a new array, and then all the chunks `concat`enated and `apply`ed. Assuming function calls usually get only a few arguments (even after spread), I think it makes more sense to accumulate them incrementally into a single temporary array; that is use a sequence of `temp.push` and `append(temp, iterable)` (new helper) calls.

```
call-spread-with-arguments baseline: 9_219 ops/sec ±1.86% (0.108ms)
call-spread-with-arguments current: 21_542 ops/sec ±0.44% (0.046ms)
```

<details>
<summary>benchmark transform</summary>

```js
// call-spread-with-arguments baseline
function foo() {
  return arguments.length;
}

function bar() {
  return foo.apply(void 0, ["ham", "spam"].concat(Array.prototype.slice.call(arguments), ["ham", "spam"]));
}

function work() {
  return bar("bacon", "egg", "spam");
}
```
```js
// call-spread-with-arguments current
function foo() {
  return arguments.length;
}

function bar() {
  var _sprd;

  return foo.apply(void 0, ((_sprd = ["ham", "spam"]).push.apply(_sprd, arguments), _sprd.push("ham", "spam"), _sprd));
}

function work() {
  return bar("bacon", "egg", "spam");
}
```
</details>

### 5. arrays with multiple spread elements
Previously, in array expression like `[x, ...a, y, ...b]` both `a` and `b` were each copied into their own array without holes, and all the pieces then concatenated. However, if there are no holes in the outer array expression, we don't need to copy `a` and `b` separately if they are already arrays, but instead use a `concat`-like helper that fills holes.

```
array-spread-multiple baseline: 3_042 ops/sec ±0.52% (0.329ms)
array-spread-multiple current: 5_618 ops/sec ±0.26% (0.178ms)
```

<details>
<summary>benchmark transform</summary>

```js
// array-spread-multiple baseline
function _toConsumableArray(arr) { return _arrayWithoutHoles(arr) || _iterableToArray(arr) || _unsupportedIterableToArray(arr) || _nonIterableSpread(); }

function _nonIterableSpread() { throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }

function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(o); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }

function _iterableToArray(iter) { if (typeof Symbol !== "undefined" && iter[Symbol.iterator] != null || iter["@@iterator"] != null) return Array.from(iter); }

function _arrayWithoutHoles(arr) { if (Array.isArray(arr)) return _arrayLikeToArray(arr); }

function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) arr2[i] = arr[i]; return arr2; }

var bar = ["egg",, "spam"];

function work(a = bar) {
  return [].concat(_toConsumableArray(a), _toConsumableArray(a), _toConsumableArray(a));
}
```
```js
// array-spread-multiple current
function _concatArrayLike() { for (var ai = 0, an = arguments.length, dn = 0; ai < an;) dn += arguments[ai++].length; var dest = new Array(dn), di = 0; for (ai = 0; ai < an;) for (var src = arguments[ai++], si = 0, sn = src.length; si < sn;) dest[di++] = src[si++]; return dest; }

function _spreadCoerceToArray(src) { return Array.isArray(src) ? src : _iterableToArray(src) || _unsupportedIterableToArray(src) || _nonIterableSpread(); }

function _nonIterableSpread() { throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }

function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(o); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }

function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) arr2[i] = arr[i]; return arr2; }

function _iterableToArray(iter) { if (typeof Symbol !== "undefined" && iter[Symbol.iterator] != null || iter["@@iterator"] != null) return Array.from(iter); }

var bar = ["egg",, "spam"];

function work(a = bar) {
  return _concatArrayLike(_spreadCoerceToArray(a), _spreadCoerceToArray(a), _spreadCoerceToArray(a));
}
```
</details>

```
array-spread-large-array baseline: 101 ops/sec ±1.39% (9.915ms)
array-spread-large-array current: 188 ops/sec ±0.32% (5.326ms)
```

<details>
<summary>benchmark transform</summary>

```js
// array-spread-large-array baseline
function _toConsumableArray(arr) { return _arrayWithoutHoles(arr) || _iterableToArray(arr) || _unsupportedIterableToArray(arr) || _nonIterableSpread(); }

function _nonIterableSpread() { throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }

function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(o); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }

function _iterableToArray(iter) { if (typeof Symbol !== "undefined" && iter[Symbol.iterator] != null || iter["@@iterator"] != null) return Array.from(iter); }

function _arrayWithoutHoles(arr) { if (Array.isArray(arr)) return _arrayLikeToArray(arr); }

function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) arr2[i] = arr[i]; return arr2; }

var bar = Array.from({
  length: 456000
}, (_, i) => i);

function foo(a) {
  return [0].concat(_toConsumableArray(a), [1], _toConsumableArray(a), [2]);
}

function work() {
  return foo(bar).length;
}
```
```js
// array-spread-large-array current
function _concatArrayLike() { for (var ai = 0, an = arguments.length, dn = 0; ai < an;) dn += arguments[ai++].length; var dest = new Array(dn), di = 0; for (ai = 0; ai < an;) for (var src = arguments[ai++], si = 0, sn = src.length; si < sn;) dest[di++] = src[si++]; return dest; }

function _spreadCoerceToArray(src) { return Array.isArray(src) ? src : _iterableToArray(src) || _unsupportedIterableToArray(src) || _nonIterableSpread(); }

function _nonIterableSpread() { throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }

function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(o); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }

function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) arr2[i] = arr[i]; return arr2; }

function _iterableToArray(iter) { if (typeof Symbol !== "undefined" && iter[Symbol.iterator] != null || iter["@@iterator"] != null) return Array.from(iter); }

var bar = Array.from({
  length: 456000
}, (_, i) => i);

function foo(a) {
  return _concatArrayLike([0], _spreadCoerceToArray(a), [1], _spreadCoerceToArray(a), [2]);
}

function work() {
  return foo(bar).length;
}
```
</details>

There's a caveat, though. If none of the spread elements are arrays, the new transform does the same amount of work as the old one — copy each iterable into an array, and then concatenate those — but the old transform uses built-in `concat`, whereas the new transform uses `concatArrayLike` helper, which is inevitably slower.

```
array-spread-large-set baseline: 93.68 ops/sec ±1.29% (11ms)
array-spread-large-set current: 86.81 ops/sec ±0.37% (12ms)
```

<details>
<summary>benchmark transform</summary>

```js
// array-spread-large-set baseline
function _toConsumableArray(arr) { return _arrayWithoutHoles(arr) || _iterableToArray(arr) || _unsupportedIterableToArray(arr) || _nonIterableSpread(); }

function _nonIterableSpread() { throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }

function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(o); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }

function _iterableToArray(iter) { if (typeof Symbol !== "undefined" && iter[Symbol.iterator] != null || iter["@@iterator"] != null) return Array.from(iter); }

function _arrayWithoutHoles(arr) { if (Array.isArray(arr)) return _arrayLikeToArray(arr); }

function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) arr2[i] = arr[i]; return arr2; }

var bar = Array.from({
  length: 456000
}, (_, i) => i);
var set = new Set(bar);

function foo(a) {
  return [0].concat(_toConsumableArray(a), [1], _toConsumableArray(a), [2]);
}

function work() {
  return foo(set).length;
}
```
```js
// array-spread-large-set current
function _concatArrayLike() { for (var ai = 0, an = arguments.length, dn = 0; ai < an;) dn += arguments[ai++].length; var dest = new Array(dn), di = 0; for (ai = 0; ai < an;) for (var src = arguments[ai++], si = 0, sn = src.length; si < sn;) dest[di++] = src[si++]; return dest; }

function _spreadCoerceToArray(src) { return Array.isArray(src) ? src : _iterableToArray(src) || _unsupportedIterableToArray(src) || _nonIterableSpread(); }

function _nonIterableSpread() { throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }

function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(o); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }

function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) arr2[i] = arr[i]; return arr2; }

function _iterableToArray(iter) { if (typeof Symbol !== "undefined" && iter[Symbol.iterator] != null || iter["@@iterator"] != null) return Array.from(iter); }

var bar = Array.from({
  length: 456000
}, (_, i) => i);
var set = new Set(bar);

function foo(a) {
  return _concatArrayLike([0], _spreadCoerceToArray(a), [1], _spreadCoerceToArray(a), [2]);
}

function work() {
  return foo(set).length;
}
```
</details>

```
array-spread-small-set baseline: 2_212 ops/sec ±0.52% (0.452ms)
array-spread-small-set current: 2_183 ops/sec ±0.8% (0.458ms)
```

<details>
<summary>benchmark transform</summary>

```js
// array-spread-small-set baseline
function _toConsumableArray(arr) { return _arrayWithoutHoles(arr) || _iterableToArray(arr) || _unsupportedIterableToArray(arr) || _nonIterableSpread(); }

function _nonIterableSpread() { throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }

function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(o); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }

function _iterableToArray(iter) { if (typeof Symbol !== "undefined" && iter[Symbol.iterator] != null || iter["@@iterator"] != null) return Array.from(iter); }

function _arrayWithoutHoles(arr) { if (Array.isArray(arr)) return _arrayLikeToArray(arr); }

function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) arr2[i] = arr[i]; return arr2; }

var bar = Array.from({
  length: 123
}, (_, i) => i);
var set = new Set(bar);

function foo(a) {
  return [0].concat(_toConsumableArray(a), [1], _toConsumableArray(a), [2]);
}

function work() {
  return foo(set).length;
}
```
```js
// array-spread-small-set current
function _concatArrayLike() { for (var ai = 0, an = arguments.length, dn = 0; ai < an;) dn += arguments[ai++].length; var dest = new Array(dn), di = 0; for (ai = 0; ai < an;) for (var src = arguments[ai++], si = 0, sn = src.length; si < sn;) dest[di++] = src[si++]; return dest; }

…