@lightmare 2021-12-02
TODO notes for self:

- [ ] in call spread, when all the parts are guaranteed array-like (such as in [this test case](https://github.com/babel/babel/pull/14015/files#diff-1f0b4c20194935c63aa476852b7e6c908615c3525678d24c596de72d48e87bcb)), I should keep using the built-in `[].concat` (holes don't matter to `f.apply`), rather than introduce helpers.